Review pass by Mirja
This review also contains feedback from our ATSSS experts: some edits are made to consider the fact that ATSSSv1 does not only have the MPTCP based solution but also ATSSS-LL (without going into any details of that as it is not relevant here). Further, only the rules to the UE are called ATSSS rules while the rules to the UPF are called PFCP rules. Therefore I removed the "ATSSS" and only talk about "rules" now. We could further clarify this by adding the following text but I don't think this is necessarily needed:

"Note that, as per 3GPP definition, the rules provided to the UE are called ATSSS Rules, while the ATSSS information provided to the UPF is carried via Multi-Access Rules, but for simplicity this document uses the term “ATSSS rules” for the ATSSS information provided to both UE and UPF."